Counter and Cleaning
A preliminary version of the Warrior Counter action has been
implemented! This needs serious work; we need to move counter out of
general defence methods and into Warrior-specific methods. While we’re
at it, we need to think about how we plan on implementing `intercept`
for the Warrior, too.

Logging has been added in this revision! The Auditor package (and
accompanying class) will keep tabs on everything we’ve done and spit
out text/a log file. This is super basic right now and needs a)
refactoring, b) rethinking, and c) extending, but it works for our
purposes for now.

Lastly, the beginnings of cleaning up the output as per Zak’s comments
on the first run-through of the game are being implemented. We’ve got
the program pausing at each attack now; see `Attack.enact()` for more.
(Line 45?)