Intercepting and Logging
Intercepting is implemented, but is broken. To observe its behaviour,
two gomezes are implemented. Also, while we’re messing with combat
skills, a Counter() class has been added to make for cleaner dealing of
this particular combat skill. The same may or may not be needed for
Intercepting, depending on mechanics. Zak must be consulted on this.

In addition, the logger got a massive revamp. Line count has more than
doubled type revamp. There are now functions for adding messages and
lines to the log, printing messages and lines, and doing both to either
too. In addition, for better logging of user input, this has been moved
to Auditor in a small way; it now has a (broken) function to take a
message explaining a boolean choice to the user, take in a Y/N and
return whether the user wants something or not.

The code has grown significantly and new features were added, as well
as parts being rewritten. I’ve pushed this new version for this reason.
However, because so many things are actually broken this time round, a
todo.txt has been added.